Fixes #7 ray-optics calculations failing with box geometry.
Box now implements its own `is_on_surface` and `normal` functions.
Prior to this release it inherited this functionality from Mesh.
The default distance tolerance EPS_ZERO seems to be a bit too small
to work well with trimesh. The problem here was that when box
needed a normal it would delegate to trimesh. We would do an
additional check that the points was on surface to a distance
tolerance of EPS_ZERO. In some cases this would fail.

We will probably see this error again when tracing pure mesh
shapes.