Cylinders have axis aligned with Y instead of Z
Meshcat has cylinder axis aligned w/ Y-axis (because that's what Three.js does).
There is no way to change that behaviour.
The easiest thing is to follow their standard in the tracer as well.

Previously, a transformation was applied to correct for this discrepancy, but
geometry that are child of others (therefore whose coordinate system is relative to the
parent) give problems because when the rotation to correct axis is applied the wrong
result might be obtained if the same correction has been applied to a parent object,
since the axis get rotated more than once.