Fixes #26 nested cylinders with Meshcat visualiser
pvtrace uses XYZ basis for definition of a cylinder but meshcat
uses XZY. This requires applying a transform to render correctly.
It was easier to do this by transforming to the world node than
use the parent/child approach. This has the consequence that in
Meshcat UI the listed scene objects are not nested, but they do
at least display correctly with repsect to rays.

Also added some options to display wireframe rather than solid
objects and some option style parameters can be set as kwargs
to the render object.