Add radiative lifetime
Luminophores can be created with an optional `lifetime` keyword.

Ray now includes a `duration` parameter.

Ray.propagate has been updated to calculate the propagation time
over the distance, this updates the `duration` attribute.

When a luminophore emits, a single exponential distribution is
sampled using the `lifetime` as the decay constant to get a
time delay for fluorescence.

Problems
--------

1. Calculating the propagation distance requires that the distance
   units are known. Up until now, pvtrace has been agnostic to
   units of the length dimension. In the commit, we assume units
   are centimetres. We will return to this issue.

2. The non-radiative time is not included. Due to the way pvtrace
   has been coded there is not an easy way to include
   non-radiative absorption lifetime. We will need to add an
   Component.absorb(ray) method which uses the non-radiative
   lifetime parameter.