Add pvtrace validation notebook
The notebook includes comparison of other LSC models:
 * ICL Raytrace Model
 * ICL 3D Flux Model
 * ECN Raytrace Model

The notebook is standalone and will generate a plot which
compares pvtrace results against the other models.

I also include Validation.png which contains the results of
a long running simulation (1000 rays, 40 runs) which brings
the uncertainty does to below +- 1%, and includes errors bars.

Fixes #37