Refactored scene.simulate to allow queue argument
The optional queue keyword argument changes the return value of
the simulate method. Without passing a queue a list of rays and
events are returned. With passing in a queue data is instead
put into the queue and nothing is returned when the function
completes.

This keeps less objects in memory provided the queue is being
monitored and items being taken out of the queue!

This is not really intended to be used by the average user, it has
been added to support multiple processing with the cli. See
changes in the cli main function.

Added progress animation.