Add support for Py3.8 on Windows (#51)
Thanks to dcambie for these changes.

* Include 3.8 in supported python version

Thanks to Meshcat>0.1.1 supporting Py38+
Actually, pvtrace also works well with Python 3.9 on linux now but there are no xxhash wheels pre-compiled for Windows (issue #43 on python-xxhash) nor it is available on conda so it needs to be compiled...

* Add py 3.8 in CI

* Replace np.float with float and np.int with int

Deprecated in NumPy 1.20:
DeprecationWarning: `np.float` is a deprecated alias for the builtin `float`. To silence this warning, use `float` by itself. Doing this will not modify any behavior and is safe.

* Remove invalid references in is_ahead()

I guess remnants of a refactoring.
According to PyCharm inspection, the function has no use in the project anyway :)

* Use augmented assignment

might be marginally faster :D

* Remove trailing spaces

* Typos

* Various style fix

* Fix some np.bool to bool comparison

also suppress pytest warning on invalid escape sequence with raw string

* Rewrite `assert bool is True` as `assert bool`