Move visualization properties from visualizer to Nodes
For many PvTrace simulation it would be nice to be able to add node visualization properties,
such as color transparency etc.
However, currently, the only visualization options (wireframe, transparency, opacity and reflectivity)
are attributes of MeshcatRenderer and apply in the same way to all the scene objects.

This can be modified at runtime *by setting the corresponding Meshcat material attributes)
but there is currently no way to save the visualization options in the scene definition (e.g. in
the YML file)

This commit allows for more granular control on the visual appearances of the scene, when rendered.
At the same time, the old global switches at the MeshcatRenderer are kept and are given priority as,
e.g. it may be needed/necessary/useful to show all scene nodes as wireframe for debugging purposes.