:bug: Force pandas read_csv to float
The example file `mock-spectrum.csv` was causing warning for a division error.
This was caused by the dtype of the loaded spectrum being int and thus havint the division [1] / 0.1
resulting in 0 instead of 0.1 when the coefficient of the absorber was 0.1 and all y values after normalization
were 1. (This was only happening with absorber and not scatterer cause scatterer coefficient was 1)