Use ProcessPoolExecutor instead of multiprocessing.Pool for multithreaded code
Also part of the standard library, is a higher level of abstraction and it internally implements
a queue so that the queue-related code can be removed.

Known bug: If workers are not set, the platform is windows and the local PC has more than 60 CPUs
a Value_error() is raised. See ProcessPoolExecutor._max_workers.
The eventuality of this corner case is remote enough that can be ignored.